dev version bump
This version 0.1.5 is not intended for Hackage release
and will not receive a git tag.

From now on, releases ending in odd numbers are versions in development.

This prevents someone from acidentally releasing a version v0.1.4
by checking out code a few commits after v0.1.4.

During the lifetime of v0.1.5,
the code starts exactly the same as v0.1.4
and ends exactly the same as v0.1.6.